refactor : "util Date -> sql.Date"
