+++ src/main/java/com/example/dailyschedule/schedule/dto/ScheduleDto.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.*;
 
-import java.util.Date;
+import java.sql.Date;
 
 
 @Getter
